Combining and pulling members up to the superclass.
