+++ app/src/main/java/be/scri/services/ItalianKeyboardIME.kt
@@ -1,39 +1,17 @@
 package be.scri.services
 
-import android.content.Context
-import android.content.res.Configuration
-import android.graphics.Color
 import android.text.InputType
 import android.util.Log
 import android.view.View
-import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.EditorInfo.IME_ACTION_NONE
 import be.scri.R
 import be.scri.databinding.KeyboardViewCommandOptionsBinding
-import be.scri.databinding.KeyboardViewKeyboardBinding
 import be.scri.helpers.MyKeyboard
-import be.scri.services.EnglishKeyboardIME.ScribeState
 import be.scri.views.MyKeyboardView
 
-class ItalianKeyboardIME : SimpleKeyboardIME() {
+class ItalianKeyboardIME : SimpleKeyboardIME("Italian") {
     override fun getKeyboardLayoutXML(): Int = R.xml.keys_letters_italian
 
-    enum class ScribeState {
-        IDLE,
-        SELECT_COMMAND,
-        TRANSLATE,
-        CONJUGATE,
-        PLURAL,
-        SELECT_VERB_CONJUNCTION,
-        SELECT_CASE_DECLENSION,
-        ALREADY_PLURAL,
-        INVALID,
-        DISPLAY_INFORMATION,
-    }
-
-    private var isAutoSuggestEnabled: Boolean = false
-    private var currentState: ScribeState = ScribeState.IDLE
-    private lateinit var keyboardBinding: KeyboardViewKeyboardBinding
     override lateinit var binding: KeyboardViewCommandOptionsBinding
     override var keyboardView: MyKeyboardView? = null
     override var keyboard: MyKeyboard? = null
@@ -48,41 +26,6 @@ class ItalianKeyboardIME : SimpleKeyboardIME() {
     override var switchToLetters = false
     override var hasTextBeforeCursor = false
 
-    override fun onInitializeInterface() {
-        super.onInitializeInterface()
-        keyboard = MyKeyboard(this, getKeyboardLayoutXML(), enterKeyType)
-    }
-
-    override fun onStartInputView(
-        editorInfo: EditorInfo?,
-        restarting: Boolean,
-    ) {
-        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
-        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
-        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
-        updateEnterKeyColor(isUserDarkMode)
-        initializeEmojiButtons()
-        isAutoSuggestEnabled = sharedPref.getBoolean("emoji_suggestions_Italian", true)
-        updateButtonVisibility(isAutoSuggestEnabled)
-        setupIdleView()
-        super.onStartInputView(editorInfo, restarting)
-        setupCommandBarTheme(binding)
-    }
-
-    private fun shouldCommitPeriodAfterSpace(language: String): Boolean {
-        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
-        return sharedPref.getBoolean("period_on_double_tap_$language", false)
-    }
-
-    override fun commitPeriodAfterSpace() {
-        if (shouldCommitPeriodAfterSpace("Italian")) {
-            val inputConnection = currentInputConnection ?: return
-            inputConnection.deleteSurroundingText(1, 0)
-            inputConnection.commitText(". ", 1)
-        }
-    }
-
     override fun onCreateInputView(): View {
         binding = KeyboardViewCommandOptionsBinding.inflate(layoutInflater)
         val keyboardHolder = binding.root
@@ -95,49 +38,6 @@ class ItalianKeyboardIME : SimpleKeyboardIME() {
         initializeEmojiButtons()
         updateUI()
         return keyboardHolder
-    }
-
-    private fun setupIdleView() {
-        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
-        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
-        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
-        when (isUserDarkMode) {
-            true -> {
-                binding.translateBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.conjugateBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.pluralBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.translateBtn.setTextColor(Color.WHITE)
-                binding.conjugateBtn.setTextColor(Color.WHITE)
-                binding.pluralBtn.setTextColor(Color.WHITE)
-                binding.separator2.setBackgroundColor(getColor(R.color.special_key_dark))
-                binding.separator3.setBackgroundColor(getColor(R.color.special_key_dark))
-            }
-            else -> {
-                binding.translateBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.conjugateBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.pluralBtn.setBackgroundColor(getColor(R.color.transparent))
-                binding.translateBtn.setTextColor(Color.BLACK)
-                binding.conjugateBtn.setTextColor(Color.BLACK)
-                binding.pluralBtn.setTextColor(Color.BLACK)
-                binding.separator2.setBackgroundColor(getColor(R.color.special_key_light))
-                binding.separator3.setBackgroundColor(getColor(R.color.special_key_light))
-            }
-        }
-
-        setupCommandBarTheme(binding)
-        binding.translateBtn.text = "Suggestion"
-        binding.conjugateBtn.text = "Suggestion"
-        binding.pluralBtn.text = "Suggestion"
-        binding.separator2.visibility = View.VISIBLE
-        binding.separator3.visibility = View.VISIBLE
-        binding.scribeKey.setOnClickListener {
-            currentState = ScribeState.SELECT_COMMAND
-            updateButtonVisibility(false)
-            Log.i("MY-TAG", "SELECT COMMAND STATE")
-            binding.scribeKey.foreground = getDrawable(R.drawable.close)
-            updateUI()
-        }
     }
 
     override fun onKey(code: Int) {
@@ -190,100 +90,10 @@ class ItalianKeyboardIME : SimpleKeyboardIME() {
         }
     }
 
-    private fun setupSelectCommandView() {
-        binding.translateBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
-        binding.conjugateBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
-        binding.pluralBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
-        setupCommandBarTheme(binding)
-        binding.translateBtn.text = "Translate"
-        binding.conjugateBtn.text = "Conjugate"
-        binding.pluralBtn.text = "Plural"
-        binding.separator2.visibility = View.GONE
-        binding.separator3.visibility = View.GONE
-        binding.scribeKey.setOnClickListener {
-            currentState = ScribeState.IDLE
-            Log.i("MY-TAG", "IDLE STATE")
-            binding.scribeKey.foreground = getDrawable(R.drawable.ic_scribe_icon_vector)
-            updateUI()
-        }
-        binding.translateBtn.setOnClickListener {
-            currentState = ScribeState.TRANSLATE
-            Log.i("MY-TAG", "TRANSLATE STATE")
-            updateUI()
-        }
-        binding.conjugateBtn.setOnClickListener {
-            Log.i("MY-TAG", "CONJUGATE STATE")
-            currentState = ScribeState.CONJUGATE
-            updateUI()
-        }
-        binding.pluralBtn.setOnClickListener {
-            Log.i("MY-TAG", "PLURAL STATE")
-            currentState = ScribeState.PLURAL
-            updateUI()
-        }
-    }
-
-    private fun switchToToolBar() {
-        val keyboardBinding = KeyboardViewKeyboardBinding.inflate(layoutInflater)
-        this.keyboardBinding = keyboardBinding
-        val keyboardHolder = keyboardBinding.root
-        keyboardView = keyboardBinding.keyboardView
-        keyboardView!!.setKeyboard(keyboard!!)
-        super.setupToolBarTheme(keyboardBinding)
-        keyboardView!!.mOnKeyboardActionListener = this
-        keyboardBinding.scribeKey.setOnClickListener {
-            currentState = ScribeState.IDLE
-            switchToCommandToolBar()
-            updateUI()
-        }
-        setInputView(keyboardHolder)
-    }
-
-    private fun switchToCommandToolBar() {
-        val binding = KeyboardViewCommandOptionsBinding.inflate(layoutInflater)
-        this.binding = binding
-        val keyboardHolder = binding.root
-        setupCommandBarTheme(binding)
-        keyboardView = binding.keyboardView
-        keyboardView!!.setKeyboard(keyboard!!)
-        keyboardView!!.mOnKeyboardActionListener = this
-        keyboardBinding.scribeKey.setOnClickListener {
-            currentState = ScribeState.IDLE
-            setupSelectCommandView()
-            updateUI()
-        }
-        setInputView(keyboardHolder)
-    }
-
-    private fun updateEnterKeyColor(isDarkMode: Boolean? = null) {
-        when (currentState) {
-            ScribeState.IDLE -> keyboardView?.setEnterKeyColor(null, isDarkMode = isDarkMode)
-            ScribeState.SELECT_COMMAND -> keyboardView?.setEnterKeyColor(null, isDarkMode = isDarkMode)
-            else -> keyboardView?.setEnterKeyColor(getColor(R.color.dark_scribe_blue))
-        }
-    }
-
     override fun onCreate() {
         super.onCreate()
         keyboard = MyKeyboard(this, getKeyboardLayoutXML(), enterKeyType)
         onCreateInputView()
         setupCommandBarTheme(binding)
-    }
-
-    private fun updateUI() {
-        when (currentState) {
-            ScribeState.IDLE -> {
-                setupIdleView()
-                initializeEmojiButtons()
-                updateButtonVisibility(isAutoSuggestEnabled)
-            }
-            ScribeState.SELECT_COMMAND -> setupSelectCommandView()
-            else -> switchToToolBar()
-        }
-        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
-        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
-        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
-        updateEnterKeyColor(isUserDarkMode)
     }
 }
