+++ app/src/main/java/be/scri/services/SimpleKeyboardIME.kt
@@ -2,6 +2,7 @@ package be.scri.services
 
 import android.content.Context
 import android.content.res.Configuration
+import android.graphics.Color
 import android.inputmethodservice.InputMethodService
 import android.text.InputType
 import android.text.InputType.TYPE_CLASS_DATETIME
@@ -9,6 +10,7 @@ import android.text.InputType.TYPE_CLASS_NUMBER
 import android.text.InputType.TYPE_CLASS_PHONE
 import android.text.InputType.TYPE_MASK_CLASS
 import android.text.TextUtils
+import android.util.Log
 import android.view.KeyEvent
 import android.view.View
 import android.view.inputmethod.EditorInfo
@@ -27,9 +29,10 @@ import be.scri.helpers.SHIFT_ON_PERMANENT
 import be.scri.views.MyKeyboardView
 // based on https://www.androidauthority.com/lets-build-custom-keyboard-android-832362/
 
-abstract class SimpleKeyboardIME :
+abstract class SimpleKeyboardIME(var language: String) :
     InputMethodService(),
     MyKeyboardView.OnKeyboardActionListener {
+
     abstract fun getKeyboardLayoutXML(): Int
 
     abstract var shiftPermToggleSpeed: Int // how quickly do we have to doubletap shift to enable permanent caps lock
@@ -56,7 +59,196 @@ abstract class SimpleKeyboardIME :
     private var emojiBtnTablet2: Button? = null
     private var emojiSpaceTablet2: View? = null
     private var emojiBtnTablet3: Button? = null
-//    abstract var keyboardViewKeyboardBinding : KeyboardViewKeyboardBinding
+    //    abstract var keyboardViewKeyboardBinding : KeyboardViewKeyboardBinding
+
+    protected var currentState: ScribeState = ScribeState.IDLE
+    protected lateinit var keyboardBinding: KeyboardViewKeyboardBinding
+    private var isAutoSuggestEnabled: Boolean = false
+
+    enum class ScribeState {
+        IDLE,
+        SELECT_COMMAND,
+        TRANSLATE,
+        CONJUGATE,
+        PLURAL,
+        SELECT_VERB_CONJUNCTION,
+        SELECT_CASE_DECLENSION,
+        ALREADY_PLURAL,
+        INVALID,
+        DISPLAY_INFORMATION,
+    }
+
+    fun getIsAccentCharacter(): Boolean {
+        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
+        val isAccentCharacter = sharedPref.getBoolean("disable_accent_character_Swedish", false)
+        return isAccentCharacter
+    }
+
+    private fun updateEnterKeyColor(isDarkMode: Boolean? = null) {
+        when (currentState) {
+            ScribeState.IDLE -> keyboardView?.setEnterKeyColor(null, isDarkMode = isDarkMode)
+            ScribeState.SELECT_COMMAND -> keyboardView?.setEnterKeyColor(null, isDarkMode = isDarkMode)
+            else -> keyboardView?.setEnterKeyColor(getColor(R.color.dark_scribe_blue))
+        }
+    }
+
+    override fun commitPeriodAfterSpace() {
+        if (getSharedPreferences("app_preferences", Context.MODE_PRIVATE).getBoolean("period_on_double_tap_$language", true)) {
+            val inputConnection = currentInputConnection ?: return
+            inputConnection.deleteSurroundingText(1, 0)
+            inputConnection.commitText(". ", 1)
+        }
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        keyboard = MyKeyboard(this, getKeyboardLayoutXML(), enterKeyType)
+        onCreateInputView()
+        setupCommandBarTheme(binding)
+        keyboardBinding = KeyboardViewKeyboardBinding.inflate(layoutInflater)
+    }
+
+    protected fun switchToCommandToolBar() {
+        val binding = KeyboardViewCommandOptionsBinding.inflate(layoutInflater)
+        this.binding = binding
+        val keyboardHolder = binding.root
+        setupCommandBarTheme(binding)
+        keyboardView = binding.keyboardView
+        keyboardView!!.setKeyboard(keyboard!!)
+        keyboardView!!.mOnKeyboardActionListener = this
+        keyboardBinding.scribeKey.setOnClickListener {
+            currentState = ScribeState.IDLE
+            setupSelectCommandView()
+            updateUI()
+        }
+        setInputView(keyboardHolder)
+    }
+
+    fun updateUI() {
+        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
+        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
+        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
+        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
+        when (currentState) {
+            ScribeState.IDLE -> {
+                setupIdleView()
+                initializeEmojiButtons()
+                updateButtonVisibility(isAutoSuggestEnabled)
+            }
+            ScribeState.SELECT_COMMAND -> setupSelectCommandView()
+            else -> switchToToolBar()
+        }
+        updateEnterKeyColor(isUserDarkMode)
+    }
+
+    private fun switchToToolBar() {
+        this.keyboardBinding = initializeKeyboardBinding()
+        val keyboardHolder = keyboardBinding.root
+        setupToolBarTheme(keyboardBinding)
+        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
+        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
+        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
+        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
+        when (isUserDarkMode) {
+            true -> {
+                keyboardBinding.topKeyboardDivider.setBackgroundColor(getColor(R.color.special_key_dark))
+            }
+            false -> {
+                keyboardBinding.topKeyboardDivider.setBackgroundColor(getColor(R.color.special_key_light))
+            }
+        }
+        keyboardView = keyboardBinding.keyboardView
+        keyboardView!!.setKeyboard(keyboard!!)
+        keyboardView!!.mOnKeyboardActionListener = this
+        keyboardBinding.scribeKey.setOnClickListener {
+            currentState = ScribeState.IDLE
+            switchToCommandToolBar()
+            updateUI()
+        }
+        setInputView(keyboardHolder)
+    }
+
+
+    private fun setupIdleView() {
+        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
+        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
+        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
+        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
+        when (isUserDarkMode) {
+            true -> {
+                binding.translateBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.conjugateBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.pluralBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.translateBtn.setTextColor(Color.WHITE)
+                binding.conjugateBtn.setTextColor(Color.WHITE)
+                binding.pluralBtn.setTextColor(Color.WHITE)
+                binding.separator2.setBackgroundColor(getColor(R.color.special_key_dark))
+                binding.separator3.setBackgroundColor(getColor(R.color.special_key_dark))
+            }
+            else -> {
+                binding.translateBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.conjugateBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.pluralBtn.setBackgroundColor(getColor(R.color.transparent))
+                binding.translateBtn.setTextColor(Color.BLACK)
+                binding.conjugateBtn.setTextColor(Color.BLACK)
+                binding.pluralBtn.setTextColor(Color.BLACK)
+                binding.separator2.setBackgroundColor(getColor(R.color.special_key_light))
+                binding.separator3.setBackgroundColor(getColor(R.color.special_key_light))
+            }
+        }
+
+        setupCommandBarTheme(binding)
+        binding.translateBtn.text = "Suggestion"
+        binding.conjugateBtn.text = "Suggestion"
+        binding.pluralBtn.text = "Suggestion"
+        binding.separator2.visibility = View.VISIBLE
+        binding.separator3.visibility = View.VISIBLE
+        binding.scribeKey.setOnClickListener {
+            currentState = ScribeState.SELECT_COMMAND
+            updateButtonVisibility(false)
+            Log.i("MY-TAG", "SELECT COMMAND STATE")
+            binding.scribeKey.foreground = getDrawable(R.drawable.close)
+            updateUI()
+        }
+    }
+
+    private fun setupSelectCommandView() {
+        binding.translateBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
+        binding.conjugateBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
+        binding.pluralBtn.setBackgroundDrawable(getDrawable(R.drawable.button_background_rounded))
+        binding.translateBtn.text = "Translate"
+        binding.conjugateBtn.text = "Conjugate"
+        binding.pluralBtn.text = "Plural"
+        binding.separator2.visibility = View.GONE
+        binding.separator3.visibility = View.GONE
+        setupCommandBarTheme(binding)
+        binding.scribeKey.setOnClickListener {
+            currentState = ScribeState.IDLE
+            Log.i("MY-TAG", "IDLE STATE")
+            binding.scribeKey.foreground = getDrawable(R.drawable.ic_scribe_icon_vector)
+            updateUI()
+        }
+        binding.translateBtn.setOnClickListener {
+            currentState = ScribeState.TRANSLATE
+            Log.i("MY-TAG", "TRANSLATE STATE")
+            updateUI()
+        }
+        binding.conjugateBtn.setOnClickListener {
+            Log.i("MY-TAG", "CONJUGATE STATE")
+            currentState = ScribeState.CONJUGATE
+            updateUI()
+        }
+        binding.pluralBtn.setOnClickListener {
+            Log.i("MY-TAG", "PLURAL STATE")
+            currentState = ScribeState.PLURAL
+            updateUI()
+        }
+    }
+
+    private fun initializeKeyboardBinding(): KeyboardViewKeyboardBinding {
+        val keyboardBinding = KeyboardViewKeyboardBinding.inflate(layoutInflater)
+        return keyboardBinding
+    }
 
     override fun onInitializeInterface() {
         super.onInitializeInterface()
@@ -72,12 +264,6 @@ abstract class SimpleKeyboardIME :
         val trimmedText = textBeforeCursor.trim()
         val lastChar = trimmedText.lastOrNull()
         return lastChar != '.'
-    }
-
-    override fun commitPeriodAfterSpace() {
-        val inputConnection = currentInputConnection ?: return
-        inputConnection.deleteSurroundingText(1, 0)
-        inputConnection.commitText(". ", 1)
     }
 
     override fun onCreateInputView(): View {
@@ -102,7 +288,7 @@ abstract class SimpleKeyboardIME :
         emojiBtnTablet3 = binding.emojiBtnTablet3
     }
 
-    fun updateButtonVisibility(isAutoSuggestEnabled: Boolean) {
+    private fun updateButtonVisibility(isAutoSuggestEnabled: Boolean) {
         val isTablet = (resources.configuration.screenLayout and Configuration.SCREENLAYOUT_SIZE_MASK) >= Configuration.SCREENLAYOUT_SIZE_LARGE
         if (isTablet) {
             pluralBtn?.visibility = if (isAutoSuggestEnabled) View.INVISIBLE else View.VISIBLE
@@ -340,7 +526,24 @@ abstract class SimpleKeyboardIME :
         }
     }
 
-    fun setupToolBarTheme(binding: KeyboardViewKeyboardBinding) {
+    override fun onStartInputView(
+        editorInfo: EditorInfo?,
+        restarting: Boolean,
+    ) {
+        val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
+        val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
+        val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
+        val isUserDarkMode = sharedPref.getBoolean("dark_mode", isSystemDarkMode)
+        updateEnterKeyColor(isUserDarkMode)
+        initializeEmojiButtons()
+        isAutoSuggestEnabled = sharedPref.getBoolean("emoji_suggestions_$language", true)
+        updateButtonVisibility(isAutoSuggestEnabled)
+        setupIdleView()
+        super.onStartInputView(editorInfo, restarting)
+        setupCommandBarTheme(binding)
+    }
+
+    private fun setupToolBarTheme(binding: KeyboardViewKeyboardBinding) {
         val sharedPref = getSharedPreferences("app_preferences", Context.MODE_PRIVATE)
         val currentNightMode = resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
         val isSystemDarkMode = currentNightMode == Configuration.UI_MODE_NIGHT_YES
